+++ api/test/login.srv.test.ts
@@ -50,6 +50,8 @@ test('POST: api/login', async () => {
         delete res.body.token;
         assert.ok(res.body.session);
         delete res.body.session;
+        assert.ok(res.body.refresh);
+        delete res.body.refresh;
 
         assert.deepEqual(res.body, {
             access: 'admin',